# samples/web/collaboration_space/source/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "协作空间项目资料管理 API",
    "description": "校园数字展馆的项目成员、资料与完整资料包导出接口。",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "sampleSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jiejian_sample_session",
        "description": "登录后建立的本地演示会话。"
      }
    }
  },
  "security": [
    {
      "sampleSession": []
    }
  ],
  "paths": {
...
    "/api/projects/{project_id}": {
      "parameters": [
        {
          "name": "project_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "readProject",
        "summary": "查看项目详情、成员与资料",
        "x-required-roles": [
          "PROJECT_OWNER",
          "MEMBER"
        ]
      }
    },
...
  }
}